refactor: 3-3 Controller 인터페이스 추가하기
