--- tomcat/src/main/java/org/apache/coyote/http11/HttpBody.java
@@ -1,4 +1,0 @@
-package org.apache.coyote.http11;
-
-public class HttpBody {
-}
