+++ tomcat/src/main/java/org/apache/coyote/http11/request_response/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request_response;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -6,13 +6,8 @@ import java.util.List;
 import java.util.Map;
 
 public class HttpCookie {
-    private Map<String, String> cookies = new HashMap<>();
 
-    public HttpCookie(String cookieHeader) {
-        Arrays.stream(cookieHeader.split(";"))
-            .map(cookie -> cookie.split("="))
-            .forEach(keyValue -> cookies.put(keyValue[0].trim(), keyValue[1].trim()));
-    }
+    private Map<String, String> cookies = new HashMap<>();
 
     // TODO indexoutofbound 고려
     public HttpCookie(List<HttpHeader> cookieHeaders) {
