+++ tomcat/src/main/java/org/apache/coyote/http11/request_response/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request_response;
 
 import java.util.ArrayList;
 import java.util.List;
