+++ tomcat/src/main/java/org/apache/coyote/http11/request_response/HttpResponse.java
@@ -1,7 +1,9 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request_response;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import org.apache.coyote.http11.HttpStatus;
 
 public class HttpResponse {
