+++ tomcat/src/main/java/org/apache/coyote/http11/request_response/QueryParameter.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request_response;
 
 public record QueryParameter(String name, String value) {
